+++ tencentcloud/mps/v20190612/errors.go
@@ -35,6 +35,9 @@ const (
 	// 内部错误。
 	INTERNALERROR = "InternalError"
 
+	// 数据错误。
+	INTERNALERROR_ACCESSDBERROR = "InternalError.AccessDBError"
+
 	// 内部错误：生成模板 ID 失败。
 	INTERNALERROR_GENDEFINITION = "InternalError.GenDefinition"