+++ tencentcloud/ocr/v20181119/client.go
@@ -2965,6 +2965,10 @@ func NewLicensePlateOCRResponse() (response *LicensePlateOCRResponse) {
 // LicensePlateOCR
 // 本接口支持对中国大陆机动车车牌的自动定位和识别，返回地域编号和车牌号码与车牌颜色信息。
 //
+// 
+//
+// 默认接口请求频率限制：10次/秒。
+//
 // 可能返回的错误码:
 //  FAILEDOPERATION_DOWNLOADERROR = "FailedOperation.DownLoadError"
 //  FAILEDOPERATION_IMAGEDECODEFAILED = "FailedOperation.ImageDecodeFailed"
@@ -2980,6 +2984,10 @@ func (c *Client) LicensePlateOCR(request *LicensePlateOCRRequest) (response *Lic
 
 // LicensePlateOCR
 // 本接口支持对中国大陆机动车车牌的自动定位和识别，返回地域编号和车牌号码与车牌颜色信息。
+//
+// 
+//
+// 默认接口请求频率限制：10次/秒。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_DOWNLOADERROR = "FailedOperation.DownLoadError"
@@ -5833,6 +5841,10 @@ func NewWaybillOCRResponse() (response *WaybillOCRResponse) {
 // WaybillOCR
 // 本接口支持市面上主流版式电子运单的识别，包括收件人和寄件人的姓名、电话、地址以及运单号等字段，精度均处于业界领先水平，识别准确率达到99%以上。
 //
+// 
+//
+// 默认接口请求频率限制：10次/秒。
+//
 // 可能返回的错误码:
 //  FAILEDOPERATION_DOWNLOADERROR = "FailedOperation.DownLoadError"
 //  FAILEDOPERATION_EMPTYIMAGEERROR = "FailedOperation.EmptyImageError"
@@ -5852,6 +5864,10 @@ func (c *Client) WaybillOCR(request *WaybillOCRRequest) (response *WaybillOCRRes
 
 // WaybillOCR
 // 本接口支持市面上主流版式电子运单的识别，包括收件人和寄件人的姓名、电话、地址以及运单号等字段，精度均处于业界领先水平，识别准确率达到99%以上。
+//
+// 
+//
+// 默认接口请求频率限制：10次/秒。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_DOWNLOADERROR = "FailedOperation.DownLoadError"