+++ tencentcloud/waf/v20180125/models.go
@@ -2493,6 +2493,57 @@ type FraudPkg struct {
 	UsedNum *int64 `json:"UsedNum,omitempty" name:"UsedNum"`
 }
 
+// Predefined struct for user
+type GetAttackDownloadRecordsRequestParams struct {
+
+}
+
+type GetAttackDownloadRecordsRequest struct {
+	*tchttp.BaseRequest
+	
+}
+
+func (r *GetAttackDownloadRecordsRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *GetAttackDownloadRecordsRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "GetAttackDownloadRecordsRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type GetAttackDownloadRecordsResponseParams struct {
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type GetAttackDownloadRecordsResponse struct {
+	*tchttp.BaseResponse
+	Response *GetAttackDownloadRecordsResponseParams `json:"Response"`
+}
+
+func (r *GetAttackDownloadRecordsResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *GetAttackDownloadRecordsResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
 type InstanceInfo struct {
 	// id
 	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`